added confirmation alert before deleting all highscores

+++ src/sample/HighscoreController.java
@@ -6,11 +6,15 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListView;
 import javafx.stage.Stage;
 
 import javafx.event.ActionEvent;
 import java.io.IOException;
+import java.util.Optional;
 
 import static javafx.scene.control.PopupControl.USE_COMPUTED_SIZE;
 
@@ -44,7 +48,16 @@ public class HighscoreController {
     }
 
     public void deleteHighScoresAction(ActionEvent actionEvent) {
-        dao.deleteAllHighscores();
-        hsListView.getItems().clear();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Upozorenje");
+        alert.setHeaderText("Jeste li sigurni da želite obrisati sve rezultate?");
+        ((Button)alert.getDialogPane().lookupButton(ButtonType.OK)).setText("Da");
+        ((Button)alert.getDialogPane().lookupButton(ButtonType.CANCEL)).setText("Odustani");
+        Optional<ButtonType> result = alert.showAndWait();
+
+        if (result.isPresent() && result.get() == ButtonType.OK) {
+            dao.deleteAllHighscores();
+            hsListView.getItems().clear();
+        }
     }
 }
